Register threads in bot and move timers to its own plugin
